pyproject.toml
[project]
name = "trmnl-dashboard"
version = "0.1.0"
description = "Flask web application to display TRMNL terminal content"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]
